requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic[email]
pydantic-settings
python-dotenv
alembic
bcrypt
python-jose
aiosmtplib
python-socketio
cryptography
# 
gemini-cli